+++ Thunderstore/BRCCustomCharactersLoader/CHANGELOG.md
@@ -1,4 +1,11 @@
 <details>
+<summary>1.0.4 </summary>
+
+* Moved BrcCustomCharacters folder from plugin folder to CharacterAPI folder inside Bepin folder.
+	* _r2modman clears entire plugin folder on each update. That means character wipe on each update. So I decided to move them to a separate folder inside Bepin. So for this version either update manually and the mod will move everything to new location or back up your characters and place them inside BepInEx\CharacterAPI\BrcCustomCharacters._
+* Added messages as to why loader can't load a character from an asset bundle.
+</details>
+<details>
 <summary>1.0.3 </summary>
 
 * Update to support BrcCustomCharacters version 2.2.2.